Settings:  App Ops
Settings: Bring back App Ops

This reverts commit 64b28a0 - uncomments the activity declaration - and
includes changes similar to those seen in 54ad5ab - adding App Ops to
the main settings screen and adding a vector drawable for the icon.

Change-Id: I65db4bdb409650a3b9d13ae6b2e2485caaf378dd

Settings: Fix an IndexOutOfBounds in App Ops

An ArrayIndexOutOfBoundsException was being thrown as 4241d0b
added a OP_MUTE_MICROPHONE op to the specific OpsTemplate without
adding a respective showPerm flag to the template. This resulted in
a mismatch of array lengths which caused trouble when calling
buildState(OpsTemplate, int, String) as that method expected the
arrays to be matching length.

This commit adds the flag as false (do not show) and adds comments to
make sure that anyone looking at the code at a later point in time can
easily see that both arrays are linked. Additionally, validation inside
the constructors of OpsTemplate have been added. This is helpful in the
case an op is removed but the respective showPerm flag is not, which
could produce unwanted behaviour by showing/hiding the wrong ops.

Change-Id: I9bb8be4cce8330d7025bfc4fa9caea71927ce425

Settings: Rewrite App Ops a bit

This commit is a short rewrite of the App Ops codebase in the Settings
application with the main goal of improving code readability and
consistency and all them good things.

Change-Id: Ia3b01bc4ec0283534350ea34975b072fa8b84847

Conflicts:
	src/com/android/settings/applications/AppOpsState.java

Settings: Sort App Ops alphabetically

The friendly comment nearby already said this should be an alphabetical
sorting mechanism. This commit makes it truly alphabetical. No messing
around with time-based sorting.

Change-Id: I6bc7cac430147d98430915e80f3e0e80fdee188a

Settings: Make App Ops show a single page

This merges all the App Ops options into a single window and fixes the
duplicate application listing bug.

Change-Id: If13c293ed81de15972f0243d79ae3d65e6aee888

Settings: Abandon the App Ops Template structure

We aren't really using the OpsTemplate anymore. Previously it was a
wrapper made to hold two constant arrays. As we are just using one
of these templates, we might as well switch to simple arrays instead.
This means we don't have to pass around any objects anywhere or worry
about which one we should use.

tl;dr Follow-up for single page merge commit.

Change-Id: I984333e60a3ab7ef31c5ea63e1c8a396bec4093a